+++ tests/specs/session-replay/mode.e2e.js
@@ -11,12 +11,12 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
     await browser.url(await browser.testHandle.assetURL('rrweb-instrumented.html', srConfig()))
       .then(() => browser.waitForSessionReplayRecording())
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       recording: true,
       initialized: true,
       events: expect.any(Array),
       mode: 1
-    }))
+    })
   })
 
   it('Full 1 Error 0 === FULL', async () => {
@@ -24,12 +24,12 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
     await browser.url(await browser.testHandle.assetURL('rrweb-instrumented.html', srConfig()))
       .then(() => browser.waitForSessionReplayRecording())
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       recording: true,
       initialized: true,
       events: expect.any(Array),
       mode: 1
-    }))
+    })
   })
 
   it('Full 0 Error 1 === ERROR', async () => {
@@ -48,12 +48,12 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
     await browser.url(await browser.testHandle.assetURL('rrweb-instrumented.html', srConfig()))
       .then(() => browser.waitForFeatureAggregate('session_replay'))
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       recording: false,
       initialized: true,
       events: [],
       mode: 0
-    }))
+    })
   })
 
   it('Full 0 Error 0 === OFF, then API called === FULL', async () => {
@@ -61,12 +61,12 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
     await browser.url(await browser.testHandle.assetURL('rrweb-instrumented.html', srConfig()))
       .then(() => browser.waitForFeatureAggregate('session_replay'))
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       recording: false,
       initialized: true,
       events: [],
       mode: 0
-    }))
+    })
 
     await Promise.all([
       browser.execute(function () {
@@ -75,12 +75,12 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
       browser.pause(1000)
     ])
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       recording: true,
       initialized: true,
       events: expect.any(Array),
       mode: 1
-    }))
+    })
   })
 
   it('Full 0 Error 1 === ERROR, then API called === FULL', async () => {
@@ -124,12 +124,12 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
     await browser.url(await browser.testHandle.assetURL('rrweb-api-pause-before-load.html', srConfig()))
       .then(() => browser.waitForSessionReplayRecording())
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       recording: true,
       initialized: true,
       events: expect.any(Array),
       mode: 1
-    }))
+    })
   })
 
   it('ERROR (seen after init) => FULL', async () => {
@@ -137,12 +137,12 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
     await browser.url(await browser.testHandle.assetURL('rrweb-instrumented.html', srConfig()))
       .then(() => browser.waitForSessionReplayRecording())
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       recording: true,
       initialized: true,
       events: expect.any(Array),
       mode: 2
-    }))
+    })
 
     await Promise.all([
       browser.execute(function () {
@@ -150,28 +150,90 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
       }), browser.pause(1000)
     ])
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       recording: true,
       initialized: true,
       events: expect.any(Array),
       mode: 1
-    }))
+    })
   })
 
-  it('ERROR (seen before init) => FULL', async () => {
+  it('ERROR (seen before init) => ERROR', async () => {
     await browser.enableSessionReplay(0, 100)
     await browser.url(await browser.testHandle.assetURL('rrweb-instrumented.html', srConfig()))
+
+    await expect(getSR()).resolves.toMatchObject({
+      recording: true,
+      initialized: true,
+      events: expect.any(Array),
+      mode: 2
+    })
+  })
+
+  it('ERROR (seen before init) --> PRELOAD => (hasReplay)', async () => {
+    await browser.enableSessionReplay(0, 100)
+    await browser.url(await browser.testHandle.assetURL('rrweb-instrumented.html', srConfig({ session_replay: { preload: true } })))
+      .then(() => browser.waitForPreloadRecorder())
       .then(() => browser.execute(function () {
         newrelic.noticeError(new Error('test'))
       }))
+    const { request: { body: err1 } } = await browser.testHandle.expectErrors()
+    const beforeLoad = err1.err[0]
+    expect(beforeLoad.params.hasReplay).toEqual(true)
+  })
+
+  it('ERROR (seen before init) --> PRELOAD but ABORTS => (!hasReplay)', async () => {
+    await browser.testHandle.clearScheduledReplies('bamServer')
+    await browser.enableSessionReplay(0, 100, 0)
+    await browser.url(await browser.testHandle.assetURL('rrweb-split-errors.html', srConfig({ session_replay: { preload: true } })))
+      .then(() => browser.waitForAgentLoad())
+    const { request: { body: err1 } } = await browser.testHandle.expectErrors()
+    const beforeLoad = err1.err[0]
+    expect(beforeLoad.params.hasReplay).toBeUndefined()
+  })
+
+  it('ERROR (seen before and after init) -- noticeError => FULL (split)', async () => {
+    await browser.enableSessionReplay(0, 100)
+    await browser.url(await browser.testHandle.assetURL('rrweb-split-errors.html', srConfig()))
       .then(() => browser.waitForSessionReplayRecording())
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
-      recording: true,
-      initialized: true,
-      events: expect.any(Array),
-      mode: 1
-    }))
+    const { request: { body: err1 } } = await browser.testHandle.expectErrors()
+
+    const beforeLoad = err1.err[0]
+    expect(beforeLoad.params.hasReplay).toBeUndefined()
+
+    const [{ request: { body: err2 } }] = await Promise.all([
+      browser.testHandle.expectErrors(),
+      browser.execute(function () {
+        newrelic.noticeError(new Error('after load'))
+      })
+    ])
+
+    const afterLoad = err2.err[0]
+    expect(afterLoad.params.hasReplay).toEqual(true)
+  })
+
+  it('ERROR (seen before and after init) -- thrown error => FULL (split)', async () => {
+    await browser.enableSessionReplay(0, 100)
+    await browser.url(await browser.testHandle.assetURL('rrweb-split-errors.html', srConfig()))
+      .then(() => browser.waitForSessionReplayRecording())
+
+    const { request: { body: err1 } } = await browser.testHandle.expectErrors()
+
+    const beforeLoad = err1.err[0]
+    expect(beforeLoad.params.hasReplay).toBeUndefined()
+
+    const [{ request: { body: err2 } }] = await Promise.all([
+      browser.testHandle.expectErrors(),
+      browser.execute(function () {
+        var scr = document.createElement('script')
+        scr.innerHTML = 'eval(\'1=2\')'
+        document.body.appendChild(scr)
+      })
+    ])
+
+    const afterLoad = err2.err[0]
+    expect(afterLoad.params.hasReplay).toEqual(true)
   })
 
   it('FULL => OFF', async () => {
@@ -179,12 +241,12 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
     await browser.url(await browser.testHandle.assetURL('rrweb-instrumented.html', srConfig()))
       .then(() => browser.waitForSessionReplayRecording())
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       recording: true,
       initialized: true,
       events: expect.any(Array),
       mode: 1
-    }))
+    })
 
     await Promise.all([
       browser.execute(function () {
@@ -192,12 +254,12 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
       }), browser.pause(1000)
     ])
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       recording: false,
       initialized: true,
       events: expect.any(Array),
       mode: 0
-    }))
+    })
   })
 
   it('blocked => OFF => API does not restart', async () => {
@@ -205,13 +267,13 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
     await browser.url(await browser.testHandle.assetURL('rrweb-instrumented.html', srConfig()))
       .then(() => browser.waitForSessionReplayRecording())
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       blocked: false,
       recording: true,
       initialized: true,
       events: expect.any(Array),
       mode: 1
-    }))
+    })
 
     await Promise.all([
       browser.execute(function () {
@@ -219,13 +281,13 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
       }), browser.pause(1000)
     ])
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       blocked: true,
       recording: false,
       initialized: true,
       events: expect.any(Array),
       mode: 0
-    }))
+    })
 
     await Promise.all([
       browser.execute(function () {
@@ -233,12 +295,12 @@ describe.withBrowsersMatching(notIE)('Session Replay Sample Mode Validation', ()
       }), browser.pause(1000)
     ])
 
-    await expect(getSR()).resolves.toEqual(expect.objectContaining({
+    await expect(getSR()).resolves.toMatchObject({
       blocked: true,
       recording: false,
       initialized: true,
       events: expect.any(Array),
       mode: 0
-    }))
+    })
   })
 })